+++ tests/fixtures/wp-package.json
@@ -1,9 +1,9 @@
 {
-  "name": "wordpress",
   "wordpress": {
+    "name": "wordpress",
     "version": "6.4.2",
     "language": "en_US",
-    "config": {
+    "WP_config": {
       "DB_NAME": "my_db_name",
       "DB_USER": "my_db_username",
       "DB_PASSWORD": "my_db_password",
